started refactoring so that we can use multiple DBMS
